# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: LMS Academy API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /courses:
    get:
      operationId: listCourses
      tags: [courses]
      summary: List courses
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/CourseStatus"
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of courses
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseListResponse"

    post:
      operationId: createCourse
      tags: [courses]
      summary: Create a new course (instructor/admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCourseRequest"
      responses:
        "201":
          description: Course created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
  /courses/catalog:
    get:
      operationId: getCourseCatalog
      tags: [courses]
      summary: Public course catalog
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Public course catalog
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseListResponse"
  /courses/{courseId}:
    get:
      operationId: getCourse
      tags: [courses]
      summary: Get course details
      security:
        - bearerAuth: []
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Course with sections and lessons
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateCourse
      tags: [courses]
      summary: Update course
      security:
        - bearerAuth: []
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCourseRequest"
      responses:
        "200":
          description: Updated course
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"

    delete:
      operationId: deleteCourse
      tags: [courses]
      summary: Delete course (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /courses/{courseId}/publish:
    post:
      operationId: publishCourse
      tags: [courses]
      summary: Publish a course
      security:
        - bearerAuth: []
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Published
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"

  # SECTIONS
  /courses/{courseId}/sections:
    get:
      operationId: listSections
      tags: [sections]
      summary: List sections for a course
      security:
        - bearerAuth: []
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Sections list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SectionWithLessons"

    post:
      operationId: createSection
      tags: [sections]
      summary: Create a section
      security:
        - bearerAuth: []
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSectionRequest"
      responses:
        "201":
          description: Section created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Section"
  /courses/{courseId}/sections/{sectionId}:
    patch:
      operationId: updateSection
      tags: [sections]
      summary: Update section
      security:
        - bearerAuth: []
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: string
        - name: sectionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSectionRequest"
      responses:
        "200":
          description: Updated section
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Section"

    delete:
      operationId: deleteSection
      tags: [sections]
      summary: Delete section
      security:
        - bearerAuth: []
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: string
        - name: sectionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # LESSONS
  /courses/{courseId}/sections/{sectionId}/lessons:
    post:
      operationId: createLesson
      tags: [lessons]
      summary: Create a lesson in a section
      security:
        - bearerAuth: []
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: string
        - name: sectionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLessonRequest"
      responses:
        "201":
          description: Lesson created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lesson"
  /courses/{courseId}/sections/{sectionId}/lessons/{lessonId}:
    get:
      operationId: getLesson
      tags: [lessons]
      summary: Get lesson details
      security:
        - bearerAuth: []
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: string
        - name: sectionId
          in: path
          required: true
          schema:
            type: string
        - name: lessonId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Lesson
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lesson"

    patch:
      operationId: updateLesson
      tags: [lessons]
      summary: Update lesson
      security:
        - bearerAuth: []
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: string
        - name: sectionId
          in: path
          required: true
          schema:
            type: string
        - name: lessonId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLessonRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lesson"

    delete:
      operationId: deleteLesson
      tags: [lessons]
      summary: Delete lesson
      security:
        - bearerAuth: []
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: string
        - name: sectionId
          in: path
          required: true
          schema:
            type: string
        - name: lessonId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ENROLLMENTS
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    # ENUMS
...
    CourseStatus:
      type: string
      enum: [draft, published, archived]
...
    LessonType:
      type: string
      enum: [video, text, quiz, live]

    # AUTH SCHEMAS
...
    Course:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        slug:
          type: string
        description:
          type: string
          nullable: true
        thumbnailUrl:
          type: string
          nullable: true
        price:
          type: number
          nullable: true
        status:
          $ref: "#/components/schemas/CourseStatus"
        instructorId:
          type: string
        instructorName:
          type: string
          nullable: true
        enrollmentCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, title, slug, status, instructorId, enrollmentCount, createdAt, updatedAt]
    CourseDetail:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        slug:
          type: string
        description:
          type: string
          nullable: true
        thumbnailUrl:
          type: string
          nullable: true
        price:
          type: number
          nullable: true
        status:
          $ref: "#/components/schemas/CourseStatus"
        instructorId:
          type: string
        instructorName:
          type: string
          nullable: true
        enrollmentCount:
          type: integer
        sections:
          type: array
          items:
            $ref: "#/components/schemas/SectionWithLessons"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, title, slug, status, instructorId, enrollmentCount, sections, createdAt, updatedAt]
    CourseListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Course"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [data, total, page, limit]
    CreateCourseRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        thumbnailUrl:
          type: string
        price:
          type: number
      required: [title]
    UpdateCourseRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        thumbnailUrl:
          type: string
        price:
          type: number
        status:
          $ref: "#/components/schemas/CourseStatus"

    # SECTION SCHEMAS
    Section:
      type: object
      properties:
        id:
          type: string
        courseId:
          type: string
        title:
          type: string
        order:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, courseId, title, order, createdAt]
    SectionWithLessons:
      type: object
      properties:
        id:
          type: string
        courseId:
          type: string
        title:
          type: string
        order:
          type: integer
        createdAt:
          type: string
          format: date-time
        lessons:
          type: array
          items:
            $ref: "#/components/schemas/Lesson"
      required: [id, courseId, title, order, createdAt, lessons]
    CreateSectionRequest:
      type: object
      properties:
        title:
          type: string
        order:
          type: integer
      required: [title]
    UpdateSectionRequest:
      type: object
      properties:
        title:
          type: string
        order:
          type: integer

    # LESSON SCHEMAS
    Lesson:
      type: object
      properties:
        id:
          type: string
        sectionId:
          type: string
        title:
          type: string
        type:
          $ref: "#/components/schemas/LessonType"
        content:
          type: string
          nullable: true
        videoUrl:
          type: string
          nullable: true
        durationMinutes:
          type: integer
          nullable: true
        order:
          type: integer
        isFree:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, sectionId, title, type, order, isFree, createdAt]
    CreateLessonRequest:
      type: object
      properties:
        title:
          type: string
        type:
          $ref: "#/components/schemas/LessonType"
        content:
          type: string
        videoUrl:
          type: string
        durationMinutes:
          type: integer
        order:
          type: integer
        isFree:
          type: boolean
      required: [title, type]
    UpdateLessonRequest:
      type: object
      properties:
        title:
          type: string
        type:
          $ref: "#/components/schemas/LessonType"
        content:
          type: string
        videoUrl:
          type: string
        durationMinutes:
          type: integer
        order:
          type: integer
        isFree:
          type: boolean

    # ENROLLMENT SCHEMAS
...