# lib/api-zod/src/generated/types/createLessonRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LMS Academy API
 * OpenAPI spec version: 0.1.0
 */
import type { LessonType } from "./lessonType";

export interface CreateLessonRequest {
  title: string;
  type: LessonType;
  content?: string;
  videoUrl?: string;
  durationMinutes?: number;
  order?: number;
  isFree?: boolean;
}