# lib/api-zod/src/generated/types/updateCourseRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LMS Academy API
 * OpenAPI spec version: 0.1.0
 */
import type { CourseStatus } from "./courseStatus";

export interface UpdateCourseRequest {
  title?: string;
  description?: string;
  thumbnailUrl?: string;
  price?: number;
  status?: CourseStatus;
}